agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
copilotkit
httpx
sse-starlette
